# Backend/openapi.json
{"openapi":"3.0.2","info":{"title":"Hotel Booking Backend","description":"OpenAPI Schema for Hotel Booking System","version":"1.0.0"},"paths":{"/users/create":{"post":{"tags":["Users"],"summary":"Create User","operationId":"create_user_users_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/get":{"get":{"tags":["Users"],"summary":"Get User","operationId":"get_user_users_get_get","parameters":[{"required":true,"schema":{"title":"Email","type":"string"},"name":"email","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/update":{"patch":{"tags":["Users"],"summary":"Update User","operationId":"update_user_users_update_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_update_user_users_update_patch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/delete":{"delete":{"tags":["Users"],"summary":"Delete User","operationId":"delete_user_users_delete_delete","parameters":[{"required":true,"schema":{"title":"Email","type":"string"},"name":"email","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/booking/create":{"post":{"tags":["Booking"],"summary":"Create Booking","operationId":"create_booking_booking_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/booking/get":{"get":{"tags":["Booking"],"summary":"Get Booking","operationId":"get_booking_booking_get_get","parameters":[{"required":true,"schema":{"title":"Bookingref","type":"string"},"name":"bookingRef","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/booking/get/all":{"get":{"tags":["Booking"],"summary":"Get Bookings","operationId":"get_bookings_booking_get_all_get","parameters":[{"required":true,"schema":{"title":"Email","type":"string"},"name":"email","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Bookings Booking Get All Get","type":"array","items":{"$ref":"#/components/schemas/BookingModel"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/booking/add":{"patch":{"tags":["Booking"],"summary":"Add Booking","operationId":"add_booking_booking_add_patch","parameters":[{"required":true,"schema":{"title":"Bookingref","type":"string"},"name":"bookingRef","in":"query"},{"required":true,"schema":{"title":"Email","type":"string"},"name":"email","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hotels":{"get":{"tags":["API"],"summary":"Get Hotels","operationId":"get_hotels_api_hotels_get","parameters":[{"required":true,"schema":{"title":"Destination","type":"string"},"name":"destination","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hotels/prices":{"get":{"tags":["API"],"summary":"Get Hotels","operationId":"get_hotels_api_hotels_prices_get","parameters":[{"required":true,"schema":{"title":"Hotelid","type":"string"},"name":"hotelId","in":"query"},{"required":true,"schema":{"title":"Destination","type":"string"},"name":"destination","in":"query"},{"required":true,"schema":{"title":"Checkindate","type":"string"},"name":"checkInDate","in":"query"},{"required":true,"schema":{"title":"Checkoutdate","type":"string"},"name":"checkOutDate","in":"query"},{"required":true,"schema":{"title":"Gueststring","type":"string"},"name":"guestString","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/prices":{"get":{"tags":["API"],"summary":"Get Hotels","operationId":"get_hotels_api_prices_get","parameters":[{"required":true,"schema":{"title":"Destination","type":"string"},"name":"destination","in":"query"},{"required":true,"schema":{"title":"Checkindate","type":"string"},"name":"checkInDate","in":"query"},{"required":true,"schema":{"title":"Checkoutdate","type":"string"},"name":"checkOutDate","in":"query"},{"required":true,"schema":{"title":"Gueststring","type":"string"},"name":"guestString","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/hello/{name}":{"get":{"summary":"Say Hello","operationId":"say_hello_hello__name__get","parameters":[{"required":true,"schema":{"title":"Name","type":"string"},"name":"name","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/schema":{"get":{"summary":"Custom Openapi","operationId":"custom_openapi_schema_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_update_user_users_update_patch":{"title":"Body_update_user_users_update_patch","required":["update","user"],"type":"object","properties":{"update":{"$ref":"#/components/schemas/UserUpdate"},"user":{"$ref":"#/components/schemas/UserAuth"}}},"BookingCreate":{"title":"BookingCreate","required":["destID","hotelID","price","bookingRef","supplierID","display","guest","payment"],"type":"object","properties":{"destID":{"title":"Destid","maxLength":4,"type":"string"},"hotelID":{"title":"Hotelid","maxLength":4,"type":"string"},"price":{"title":"Price","exclusiveMinimum":0.0,"type":"number"},"bookingRef":{"title":"Bookingref","type":"string"},"supplierID":{"title":"Supplierid","type":"string"},"supplierRes":{"title":"Supplierres","type":"array","items":{},"default":[]},"display":{"$ref":"#/components/schemas/DisplayInfo"},"guest":{"$ref":"#/components/schemas/GuestInfo"},"payment":{"$ref":"#/components/schemas/PaymentInfo"}}},"BookingModel":{"title":"BookingModel","required":["destID","hotelID","price","bookingRef","supplierID","display","guest","payment"],"type":"object","properties":{"destID":{"title":"Destid","maxLength":4,"type":"string"},"hotelID":{"title":"Hotelid","maxLength":4,"type":"string"},"price":{"title":"Price","exclusiveMinimum":0.0,"type":"number"},"bookingRef":{"title":"Bookingref","type":"string"},"supplierID":{"title":"Supplierid","type":"string"},"supplierRes":{"title":"Supplierres","type":"array","items":{},"default":[]},"display":{"$ref":"#/components/schemas/DisplayInfo"},"guest":{"$ref":"#/components/schemas/GuestInfo"},"payment":{"$ref":"#/components/schemas/PaymentInfo"},"_id":{"title":" Id","type":"string","examples":["5eb7cf5a86d9755df3a6c593","5eb7cfb05e32e07750a1756a"]}},"description":"Document Mapping class.\n\nFields:\n\n- `id` - MongoDB document ObjectID \"_id\" field.\nMapped to the PydanticObjectId class\n\nInherited from:\n\n- Pydantic BaseModel\n- [UpdateMethods](https://roman-right.github.io/beanie/api/interfaces/#aggregatemethods)"},"BookingOut":{"title":"BookingOut","required":["destID","hotelID","price","display","supplierID","supplierRes","bookingRef","guest","payment"],"type":"object","properties":{"destID":{"title":"Destid","type":"string"},"hotelID":{"title":"Hotelid","type":"string"},"price":{"title":"Price","type":"number"},"display":{"$ref":"#/components/schemas/DisplayInfo"},"supplierID":{"title":"Supplierid","type":"string"},"supplierRes":{"title":"Supplierres","type":"array","items":{}},"bookingRef":{"title":"Bookingref","type":"string"},"guest":{"$ref":"#/components/schemas/GuestInfo"},"payment":{"$ref":"#/components/schemas/PaymentInfo"}}},"DisplayInfo":{"title":"DisplayInfo","required":["numOfNights","startDate","endDate","adults","children","roomType"],"type":"object","properties":{"numOfNights":{"title":"Numofnights","type":"integer"},"startDate":{"title":"Startdate","type":"string","format":"date"},"endDate":{"title":"Enddate","type":"string","format":"date"},"adults":{"title":"Adults","type":"integer"},"children":{"title":"Children","type":"integer"},"roomType":{"title":"Roomtype","type":"array","items":{"type":"string"}},"message":{"title":"Message","type":"string"}}},"GuestInfo":{"title":"GuestInfo","required":["firstName","lastName","phone","email"],"type":"object","properties":{"salutation":{"title":"Salutation","type":"string"},"firstName":{"title":"Firstname","type":"string","format":"password","writeOnly":true},"lastName":{"title":"Lastname","type":"string","format":"password","writeOnly":true},"phone":{"title":"Phone","type":"string","format":"password","writeOnly":true},"email":{"title":"Email","type":"string","format":"email"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"PaymentInfo":{"title":"PaymentInfo","required":["paymentID","payeeID"],"type":"object","properties":{"paymentID":{"title":"Paymentid","type":"string","format":"password","writeOnly":true},"payeeID":{"title":"Payeeid","type":"string","format":"password","writeOnly":true}}},"UserAuth":{"title":"UserAuth","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","type":"string","format":"password","writeOnly":true}}},"UserCreate":{"title":"UserCreate","required":["firstName","lastName","email","password"],"type":"object","properties":{"firstName":{"title":"Firstname","type":"string"},"lastName":{"title":"Lastname","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","type":"string"}}},"UserModel":{"title":"UserModel","required":["firstName","lastName","email","password"],"type":"object","properties":{"firstName":{"title":"Firstname","type":"string"},"lastName":{"title":"Lastname","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","type":"string","format":"password","writeOnly":true},"_id":{"title":" Id","type":"string","examples":["5eb7cf5a86d9755df3a6c593","5eb7cfb05e32e07750a1756a"]},"disabled":{"title":"Disabled","type":"boolean","default":false},"bookings":{"title":"Bookings","type":"array","items":{"allOf":[{"$ref":"#/components/schemas/BookingModel"}]},"default":[]}},"description":"Document Mapping class.\n\nFields:\n\n- `id` - MongoDB document ObjectID \"_id\" field.\nMapped to the PydanticObjectId class\n\nInherited from:\n\n- Pydantic BaseModel\n- [UpdateMethods](https://roman-right.github.io/beanie/api/interfaces/#aggregatemethods)"},"UserOut":{"title":"UserOut","required":["email","bookings","disabled"],"type":"object","properties":{"firstName":{"title":"Firstname","type":"string"},"lastName":{"title":"Lastname","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"bookings":{"title":"Bookings","type":"array","items":{"allOf":[{"$ref":"#/components/schemas/BookingModel"}]}},"disabled":{"title":"Disabled","type":"boolean"}}},"UserUpdate":{"title":"UserUpdate","required":["email"],"type":"object","properties":{"firstName":{"title":"Firstname","type":"string"},"lastName":{"title":"Lastname","type":"string"},"email":{"title":"Email","type":"string","format":"email"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}